requirements.txt
Flask==2.3.2
Flask-SQLAlchemy
psycopg2-binary
prometheus_flask_exporter